Make email checking function and implement it
Made a function in the user-repository called checkIfUserExistsByEmail.
Its parameter takes a plain text email, that will be used to search
existing user that use that email. This function will be called in the
user-service.

Made a function in the user-service called checkEmailAvailability.
Using the function in controller to check email Availability. Return a
promise boolean.

In controller, implement and call the function from user-service for
createUser and updateUser to make sure that no duplicate email used.